refactor: change search bar component path (#41)
